Histogram Changed To TrendLineChart and Get Data
