+++ web/src/component/TrendLineChart/index.module.scss
@@ -6,3 +6,6 @@
   display: table-cell;
   padding: 0px 20px 0px 0px;
 }
+.linechart-wrap {
+  margin-left: 28;
+}
